feat(ui): Position UI windows by anchor and offsets
Adds four new options to customize the position of the UI popup window(s):

- `vertical_anchor`
  - Options: `"top"`, `"center"`, `"bottom"`
  - Default: `"center"`
- `horizontal_anchor`
  - Options: `"left"`, `"center"`, `"right"`
  - Default: `"center"`
- `vertical_offset`
  - Default: 0
- `horizontal_offset`
  - Default: 0

The anchor and the offset values adjust the row and column position for each window (preview, patterns, replacements, options).

If the window would be out of bounds due to a combination of the anchor or offset opts and the overall size of the windows, the adjusted row/col positions of the UI are set to a max limit, placing them at their opposite anchor point.